Move src to backup folder; rename src_complete to src
