+++ src/gain.txt
